Extract tokenizer.rb
