+++ lib/rbs/inline.rb
@@ -9,6 +9,7 @@ require "rbs"
 
 require "rbs/inline/node_utils"
 require "rbs/inline/annotation_parser"
+require "rbs/inline/annotation_parser/tokenizer"
 require "rbs/inline/ast/annotations"
 require "rbs/inline/ast/comment_lines"
 require "rbs/inline/ast/tree"
